Turns polygons into covering spaces
This takes as input a list of vertex angles and their multiplicities. This information uniquely defines the cone angles of a polygon and this program takes that information and prints out information about cone angles which uniquely characterize the canonical K-covering  for the polygon in question. Then you can do other things with this information, maybe with some of the other programs here.